refactor(web): split the three largest size-ratchet functions
Removes the last three max-lines-per-function suppressions:

- DatabaseBackupSettings' queries, mutations, clipboard and download
  helpers moved into useBackupActions, and the periodic-backup enable
  confirmation into BackupEnableConfirm.
- Arena.tsx keeps the header, effects and modals; the controls card,
  the bracket and run bar, and the response grid are ArenaControls,
  ArenaBracketBar and ArenaResponseGrid, each taking the ref-free arena
  view object plus the one ref it reads.
- Chat.tsx keeps the header, conversation config and message area; the
  controls card and the input and stats area are ChatControls and
  ChatInputArea on the same pattern.

ArenaView/ArenaRefs and ChatView/ChatRefs are exported beside their
hooks so the sections share one type with the page. No behaviour
change; the existing page tests render every moved section.
