+++ web/src/pages/Settings/DatabaseBackupSettings.tsx
@@ -1,19 +1,8 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useQueryClient } from "@tanstack/react-query";
 import { Trans, useTranslation } from "react-i18next";
-import {
-	AlertTriangle,
-	Copy,
-	Download,
-	HardDrive,
-	Plus,
-	Trash2,
-	Upload,
-} from "@/lib/icons";
+import { Copy, Download, HardDrive, Plus, Trash2, Upload } from "@/lib/icons";
 import { api, getAuthHeaders } from "../../api/client";
-import type { BackupClassification } from "../../api/types";
 import { LoadingSpinner } from "../../components/LoadingSpinner";
-import { Modal } from "../../components/Modal";
 import { RestoreConfirmModal } from "../../components/RestoreConfirmModal";
 import { SettingsGroup } from "../../components/SettingsGroup";
 import { SettingsSection } from "../../components/SettingsSection";
@@ -21,8 +10,9 @@ import { SettingsSlider } from "../../components/SettingsSlider";
 import { Spinner } from "../../components/Spinner";
 import { Toggle } from "../../components/Toggle";
 import { useToast } from "../../context/ToastContext";
-import { useCopyToClipboard } from "../../hooks/useCopyToClipboard";
 import { formatDateTimeShort } from "../../utils/format";
+import { BackupEnableConfirm } from "./BackupEnableConfirm";
+import { useBackupActions } from "./useBackupActions";
 
 interface DatabaseBackupSettingsProps {
 	collapsed: boolean;
@@ -30,7 +20,6 @@ interface DatabaseBackupSettingsProps {
 	managed?: boolean;
 }
 
-// eslint-disable-next-line max-lines-per-function -- size ratchet: split this component
 export function DatabaseBackupSettings({
 	collapsed,
 	onToggle,
@@ -39,225 +28,36 @@ export function DatabaseBackupSettings({
 	const { t } = useTranslation();
 	const { toast } = useToast();
 	const queryClient = useQueryClient();
-	const [confirmDelete, setConfirmDelete] = useState<string | null>(null);
-	const [restoreFile, setRestoreFile] = useState<File | null>(null);
-	const [showRestoreModal, setShowRestoreModal] = useState(false);
-	const [isRestoring, setIsRestoring] = useState(false);
-	const [showEnableConfirm, setShowEnableConfirm] = useState(false);
-	const [prunePreview, setPrunePreview] = useState<BackupClassification | null>(
-		null,
-	);
-	const fileInputRef = useRef<HTMLInputElement>(null);
-	const pollingRef = useRef(false);
-
-	useEffect(() => {
-		return () => {
-			pollingRef.current = false;
-		};
-	}, []);
-
-	const { data: backups, isLoading } = useQuery({
-		queryKey: ["backups"],
-		queryFn: () => api.backups.list(),
-	});
-
-	const { data: settings, isPending: settingsPending } = useQuery({
-		queryKey: ["settings"],
-		queryFn: () => api.settings.get(),
-	});
-
-	// GFS bucket per backup, so each row can carry a Grandfather/Father/Son tag.
-	// Sourced from the prune-preview classifier, which groups every backup by age
-	// against the configured retention, so the labels track the same rotation the
-	// sliders above configure.
-	//
-	// Its own key, outside the "backups" prefix: the preview is a POST the server
-	// treats as a read, and running it on every backups invalidation (create,
-	// delete, prune, restore) would re-classify an unchanged list. The key
-	// carries what can change the classification: the set of backups on disk and
-	// the retention the classifier applies. It waits for the settings so a late
-	// settings answer does not cost a second read (isPending, not undefined, so
-	// a settings error still lets the buckets load).
-	const backupNames = useMemo(
-		() => (backups ?? []).map((b) => b.filename).sort(),
-		[backups],
-	);
-	const { data: classification } = useQuery({
-		queryKey: [
-			"backup-classification",
-			backupNames,
-			settings?.backup_son_retention,
-			settings?.backup_father_retention,
-			settings?.backup_grandfather_retention,
-		],
-		queryFn: () => api.backups.prunePreview(),
-		enabled: backupNames.length > 0 && !settingsPending,
-		staleTime: Number.POSITIVE_INFINITY,
-	});
-
-	const gfsLabel = useMemo(() => {
-		const m = new Map<string, "G" | "F" | "S">();
-		for (const b of classification?.grandfather ?? []) m.set(b.filename, "G");
-		for (const b of classification?.father ?? []) m.set(b.filename, "F");
-		for (const b of classification?.son ?? []) m.set(b.filename, "S");
-		return m;
-	}, [classification]);
-
-	const createMutation = useMutation({
-		mutationFn: () => api.backups.create(),
-		onSuccess: () => {
-			queryClient.invalidateQueries({ queryKey: ["backups"] });
-		},
-		onError: (err: Error) => {
-			toast(
-				t("settings.backup.backupFailed", { message: err.message }),
-				"error",
-			);
-		},
-	});
-
-	const deleteMutation = useMutation({
-		mutationFn: (filename: string) => api.backups.delete(filename),
-		onSuccess: () => {
-			queryClient.invalidateQueries({ queryKey: ["backups"] });
-			setConfirmDelete(null);
-			toast(t("settings.backup.backupDeleted"), "success");
-		},
-		onError: (err: Error) => {
-			toast(
-				t("settings.backup.deleteFailed", { message: err.message }),
-				"error",
-			);
-		},
-	});
-
-	// Settings for periodic backup
-	const settingsUpdateMutation = useMutation({
-		mutationFn: (updates: Record<string, string>) =>
-			api.settings.update(updates),
-		onSuccess: () => {
-			queryClient.invalidateQueries({ queryKey: ["settings"] });
-		},
-		onError: (err: Error) => {
-			toast(
-				t("settings.common.failedToSave", { message: err.message }),
-				"error",
-			);
-		},
-	});
-
-	const backupEnabled = settings?.backup_enabled === "true";
-	// Parse interval: backend stores as Go duration string (e.g. "86400s" or "24h").
-	// Display and edit in hours.
-	const rawInterval = settings?.backup_interval || "24h";
-	const intervalHours = (() => {
-		const hMatch = rawInterval.match(/^(\d+(?:\.\d+)?)h$/);
-		if (hMatch) return Number(hMatch[1]);
-		const sMatch = rawInterval.match(/^(\d+(?:\.\d+)?)s$/);
-		if (sMatch) return Math.round((Number(sMatch[1]) / 3600) * 10) / 10;
-		return 24;
-	})();
-	const sonRetention = Number(settings?.backup_son_retention || "7");
-	const fatherRetention = Number(settings?.backup_father_retention || "4");
-	const grandfatherRetention = Number(
-		settings?.backup_grandfather_retention || "3",
-	);
-
-	function formatBytes(bytes: number): string {
-		if (bytes === 0) return "0 B";
-		const k = 1024;
-		const sizes = ["B", "KB", "MB", "GB", "TB"];
-		const i = Math.min(
-			Math.floor(Math.log(bytes) / Math.log(k)),
-			sizes.length - 1,
-		);
-		return `${Number.parseFloat((bytes / k ** i).toFixed(1))} ${sizes[i]}`;
-	}
-
-	// navigator.clipboard only exists in secure contexts; a dashboard served
-	// over plain HTTP on a LAN (a normal self-hosted setup) has none, so fall
-	// back to the legacy selection-based copy rather than fail on the one
-	// button that exists to unblock a verified restore.
-	const writeClipboard = useCallback(
-		async (text: string) => {
-			if (navigator.clipboard?.writeText) {
-				await navigator.clipboard.writeText(text);
-				return;
-			}
-			const holder = document.createElement("textarea");
-			holder.value = text;
-			holder.setAttribute("readonly", "");
-			holder.style.position = "fixed";
-			holder.style.opacity = "0";
-			document.body.appendChild(holder);
-			holder.select();
-			let copied: boolean;
-			try {
-				copied = document.execCommand("copy");
-			} finally {
-				document.body.removeChild(holder);
-			}
-			if (!copied) {
-				throw new Error(t("common.failedToCopy"));
-			}
-		},
-		[t],
-	);
-	// One clipboard path for the whole dashboard, with this page's fallback
-	// writer in front of it. The button reports through a toast, so the "Copied"
-	// flag is not tracked.
-	const { copy } = useCopyToClipboard({
-		write: writeClipboard,
-		trackCopied: false,
-	});
-
-	// Puts the backup's signature sidecar on the clipboard for the restore form.
-	// The download hands over the dump alone, so without shell access to the
-	// backup directory this is the only way to get the value a verified restore
-	// needs.
-	const copySignature = async (filename: string) => {
-		try {
-			const { signature } = await api.backups.signature(filename);
-			if (!(await copy(signature))) {
-				throw new Error(t("common.failedToCopy"));
-			}
-			toast(t("settings.backup.signatureCopied"), "success");
-		} catch (err) {
-			toast(
-				t("settings.backup.signatureCopyFailed", {
-					message: (err as Error).message,
-				}),
-				"error",
-			);
-		}
-	};
-
-	const downloadBackup = async (filename: string) => {
-		try {
-			const response = await fetch(api.backups.downloadUrl(filename), {
-				headers: getAuthHeaders(),
-			});
-			if (!response.ok) {
-				throw new Error(`Download failed: ${response.status}`);
-			}
-			const blob = await response.blob();
-			const url = URL.createObjectURL(blob);
-			const a = document.createElement("a");
-			a.href = url;
-			a.download = filename;
-			document.body.appendChild(a);
-			a.click();
-			document.body.removeChild(a);
-			URL.revokeObjectURL(url);
-		} catch (err) {
-			toast(
-				t("settings.backup.downloadFailed", {
-					message: (err as Error).message,
-				}),
-				"error",
-			);
-		}
-	};
+	const {
+		confirmDelete,
+		setConfirmDelete,
+		restoreFile,
+		setRestoreFile,
+		showRestoreModal,
+		setShowRestoreModal,
+		isRestoring,
+		setIsRestoring,
+		showEnableConfirm,
+		setShowEnableConfirm,
+		prunePreview,
+		setPrunePreview,
+		fileInputRef,
+		pollingRef,
+		gfsLabel,
+		createMutation,
+		deleteMutation,
+		settingsUpdateMutation,
+		backupEnabled,
+		intervalHours,
+		sonRetention,
+		fatherRetention,
+		grandfatherRetention,
+		copySignature,
+		downloadBackup,
+		backups,
+		isLoading,
+		formatBytes,
+	} = useBackupActions();
 
 	return (
 		<SettingsSection
@@ -429,85 +229,13 @@ export function DatabaseBackupSettings({
 					</div>
 				</SettingsGroup>
 
-				{/* Double-confirm modal for enabling periodic backup */}
-				{showEnableConfirm && (
-					<Modal
-						onClose={() => {
-							setShowEnableConfirm(false);
-							setPrunePreview(null);
-						}}
-						title={t("settings.backup.rotation.confirmEnableTitle")}
-						maxWidth="max-w-lg"
-					>
-						<div className="space-y-3">
-							<div className="flex items-start gap-2 text-amber-400">
-								<AlertTriangle size={18} className="shrink-0 mt-0.5" />
-								<p className="text-sm text-(--text-secondary)">
-									{t("settings.backup.rotation.confirmEnableDescription")}
-								</p>
-							</div>
-							<div className="space-y-2">
-								<p className="text-sm text-(--text-primary)">
-									{t("settings.backup.rotation.confirmEnableWouldRemove", {
-										count: prunePreview?.prune?.length ?? 0,
-									})}
-								</p>
-								<div className="max-h-40 overflow-y-auto rounded bg-(--surface-elevated) border border-(--border-default) p-2">
-									{(prunePreview?.prune ?? []).map((b) => (
-										<div
-											key={b.filename}
-											className="text-xs font-mono text-(--text-secondary) py-0.5"
-										>
-											{b.filename}
-										</div>
-									))}
-								</div>
-							</div>
-							<div className="flex justify-end gap-2 pt-2">
-								<button
-									type="button"
-									onClick={() => {
-										setShowEnableConfirm(false);
-										setPrunePreview(null);
-									}}
-									className="ui-btn ui-btn-secondary"
-								>
-									{t("common.cancel")}
-								</button>
-								<button
-									type="button"
-									onClick={async () => {
-										try {
-											if ((prunePreview?.prune?.length ?? 0) > 0) {
-												await api.backups.prune();
-											}
-											await settingsUpdateMutation.mutateAsync({
-												backup_enabled: "true",
-											});
-											toast(
-												t("settings.backup.rotation.pruneSuccess", {
-													count: prunePreview?.prune?.length ?? 0,
-												}),
-												"success",
-											);
-										} catch {
-											toast(t("settings.backup.rotation.pruneFailed"), "error");
-										} finally {
-											setShowEnableConfirm(false);
-											setPrunePreview(null);
-											queryClient.invalidateQueries({
-												queryKey: ["backups"],
-											});
-										}
-									}}
-									className="ui-btn ui-btn-primary"
-								>
-									{t("settings.backup.confirm")}
-								</button>
-							</div>
-						</div>
-					</Modal>
-				)}
+				<BackupEnableConfirm
+					showEnableConfirm={showEnableConfirm}
+					setShowEnableConfirm={setShowEnableConfirm}
+					prunePreview={prunePreview}
+					setPrunePreview={setPrunePreview}
+					settingsUpdateMutation={settingsUpdateMutation}
+				/>
 
 				{/* Restore requirements */}
 				<div className="rounded-[var(--radius-card,0.375rem)] border border-(--accent)/30 bg-(--accent)/5 p-3 space-y-2">
